+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.3.1] - 2023-04-06
 ### Added
 - Add metrics colletion for backend observability. [Read more](https://github.com/captainjapeng/watch-on-floatplane#data-collection-and-privacy)